requirements.txt
langchain-groq
streamlit
pandas
python-dotenv
